requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.22
aiopg==1.3.3
aiosqlite==0.17.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
braceexpand==0.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
discord-components==2.1.2
discordSuperUtils==0.3.0
idna==3.3
import-expression==1.1.4
importlib-metadata==4.8.2
jishaku @ git+https://github.com/Gorialis/jishaku@cf6dd32f3b2693067540b1a5247ecbb41b49e104
motor==2.5.1
multidict==5.2.0
Pillow==8.4.0
psycopg2-binary==2.9.2
py-cord==1.7.3
py-cord-components @ git+https://github.com/spacedev-official/py-cord-components.git@b98bab9eaeb6d3ed428d519506c18228035152d5
pycparser==2.21
pymongo==3.12.1
PyMySQL==0.9.3
PyNaCl==1.4.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
spotipy==2.19.0
typing-extensions==4.0.0
urllib3==1.26.7
wavelink==0.9.10
yarl==1.7.2
zipp==3.6.0
